fix: sem detail log 수정
