+++ src/common/enum/equipment.enum.ts
@@ -1,0 +1,17 @@
+export enum FormType {
+  MANIPULATOR = 'manipulator',
+  TORSO = 'torso',
+  AMR = 'amr',
+}
+
+export enum ManipulatorType {
+  LEFT = 'left',
+  RIGHT = 'right',
+}
+
+export enum AmrLogType {
+  VELOCITY = 'velocity',
+  OBSTACLE = 'obstacle',
+  DOCKING_PRECISION = 'docking_precision',
+  MOVING_PRECISION = 'moving_precision',
+}
