+++ src/common/interface/system/equipment.interface.ts
@@ -1,3 +1,25 @@
+export interface ManipulatorPositionPayload {
+  dateTime: string;
+  x: number;
+  y: number;
+  z: number;
+  rx: number;
+  ry: number;
+  rz: number;
+  base: number;
+  shoulder: number;
+  elbow: number;
+  wrist1: number;
+  wrist2: number;
+}
+
+export interface TorsoPositionPayload {
+  dateTime: string;
+  x: number;
+  z: number;
+  theta: number;
+}
+
 export interface AmrVelocityPayload {
   dateTime: string;
   x: number;
