--- src/common/interface/system/manipulator.interface.ts
@@ -1,14 +1,0 @@
-export interface ManipulatorPositionPayload {
-  dateTime: string;
-  x: number;
-  y: number;
-  z: number;
-  rx: number;
-  ry: number;
-  rz: number;
-  base: number;
-  shoulder: number;
-  elbow: number;
-  wrist1: number;
-  wrist2: number;
-}
