--- src/common/interface/system/torso.interface.ts
@@ -1,6 +1,0 @@
-export interface TorsoPositionPayload {
-  dateTime: string;
-  x: number;
-  z: number;
-  theta: number;
-}
